Upstream Watchlist Plugin
What is a Watchlist? This is a functionality some IDEs offer while
debugging code. A Watchlist is related to the display of all variables
accessible in the current scope (aka Variable Explorer in Spyder). But
there is an important distinction: A Watchlist consists of a
user-definable list of expressions. These expressions are evaluated
after a debugger step and the result of the evaluation is displayed.

Features:

* List of expressions can be edited at any time
* Expressions can be added by several means (toolbar buttons, right
  click menu, double clicking at the end of the list, drag & drop of
  text)
* Existing entries can be changed (double click)
* The value of an expression is shown in bold font if its value changes
  after a debugger step

Known issue: Watchlist expressions are not restored when a new console
is opened (internal version has a fix which must be adapted).